link subvoice to voice + aquire host information

+++ Source/SynthVoice.h
@@ -27,7 +27,7 @@ public:
     
     int trigger;
     
-    double frequency = 440;
+    float frequency = 440;
     
     // get UI data
     void getSlider(float sliderValue, String ID);
